+++ packages/web-tracing/src/index.ts
@@ -6,4 +6,6 @@ export { getDefaultOTELInstrumentations } from './getDefaultOTELInstrumentations
 
 export { TracingInstrumentation } from './instrumentation';
 
+export { getSamplingDecision } from './sampler';
+
 export type { FaroTraceExporterConfig, TracingInstrumentationOptions } from './types';